refactor bbking so it works with 1.9 app loading
