+++ arkestrator/bbking/bbtags/code.py
@@ -1,8 +1,6 @@
-from bbking import BBTag
+from bbking.tags import BBTag
 
 __all__ = ['BBCode']
 
 class BBCode(BBTag):
     tag_name = 'code'
-
-
