+++ arkestrator/bbking/bbtags/quote.py
@@ -1,5 +1,5 @@
 
-from bbking import BBTag
+from bbking.tags import BBTag
 
 __all__ = ['BBQuote']
