+++ arkestrator/bbking/bbtags/text.py
@@ -1,5 +1,5 @@
 
-from bbking import BBTag
+from bbking.tags import BBTag
 
 __all__ = ['BBTagItalic', 'BBTagBold', 'BBTagUnderline', 'BBTagStrikethrough']
