+++ arkestrator/bbking/errors.py
@@ -1,0 +1,9 @@
+class CompilationError(Exception):
+    pass
+
+class TagDoesNotExist(CompilationError):
+    pass
+
+class UnnamedTagException(CompilationError):
+    pass
+
